+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/fuhdao/preset/TriggerInfoPresetCriteriaMaker.java
@@ -52,7 +52,7 @@ public class TriggerInfoPresetCriteriaMaker implements PresetCriteriaMaker {
             if (Objects.isNull(objects[0])) {
                 detachedCriteria.add(Restrictions.isNull("pointLongId"));
             } else {
-                //noinspection unchecked
+                @SuppressWarnings("unchecked")
                 List<LongIdKey> longIdKeys = (List<LongIdKey>) objects[0];
                 if (longIdKeys.isEmpty()) {
                     detachedCriteria.add(Restrictions.isNull("longId"));